商品管理模块完善

+++ src/main/java/com/corngo/admin/service/SystemRealm.java
@@ -81,7 +81,8 @@ public class SystemRealm extends AuthorizingRealm {
             if (menu.getType().equals("1")){
                 ArrayList<SysMenu> child = new ArrayList<>();
                 for (SysMenu m:menus){
-                    if (m.getParentId() == menu.getId()){
+//                    System.out.println("parentId:"+m.getParentId() + "---mId:"+menu.getId() +"---"+(m.getParentId() == menu.getId()) + "---"+((long)m.getParentId() == (long)menu.getId()));
+                    if ((long)m.getParentId() == (long)menu.getId()){
                         child.add(m);
                     }
                 }
